requirement.txt
django-admin=3.0
python-decouple
psycopg2
Pillow